Updating classes with new logging methods and gem dependencies; adding Syndi::IRC::Client and Syndi::Logging; making the spec helper more useful.

+++ spec/helper.rb
@@ -10,6 +10,7 @@ require 'rspec/mocks'
 require 'fileutils'
 require 'stringio'
 require 'tmpdir'
+require 'celluloid/io'
     
 require 'syndi'
 
@@ -18,7 +19,50 @@ Syndi.dir = $temp_dir
 $VERBOSITY = -1
 Syndi.celluloid_log
 
+class TestServer
+  include Celluloid::IO
+  attr_accessor :server
+
+  def initialize address, port
+    puts "*** Experimental TCP server starting on #{address}:#{port}"
+    @server = TCPServer.new address, port
+    @buffer = ''
+    async.run
+  end
+
+  def run
+    loop { async.handle_conn @server.accept }
+  end
+
+  def handle_conn client
+    _, port, addr = client.peeraddr
+    puts "*** Experimental connection from #{addr}:#{port}"
+    loop { @buffer << client.readpartial(4096) }
+  rescue EOFError
+    puts "*** Connection with #{addr}:#{port} lost"
+    client.close
+  end
+
+  def finalize
+    puts '*** Server terminating'
+    @server.close if @server
+  end
+
+  def buffer
+    value = @buffer
+    @buffer = ''
+    value
+  end
+end
+
+
 module Helper
+
+  def with_tcp_serv
+    server = TestServer.new '127.0.0.1', 7884
+    yield '127.0.0.1', 7884, server
+    server.terminate
+  end
 
 end
 
@@ -27,7 +71,7 @@ RSpec.configure do |conf|
 
   conf.after(:all) do
     Dir.chdir __dir__
-    FileUtils.remove_entry $temp_dir
+    FileUtils.remove_entry $temp_dir if Dir.exists? $temp_dir
     Syndi.dir = __dir__
   end
